camí traçat

+++ android/emergps/src/upc/pxc/emergps/Dades.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 public class Dades {
 
 	String id;
-	Float posx, posy;
+	Double posx, posy;
 	String comment;
 	ArrayList<Staff> poli = new ArrayList();
 	ArrayList<Staff> bomb = new ArrayList();
@@ -14,8 +14,8 @@ public class Dades {
 	public void setDades(String s){
 		String[] dat = s.split("&");
 		id = dat[0];
-		posx = Float.parseFloat(dat[1]);
-		posy = Float.parseFloat(dat[2]);
+		posx = Double.parseDouble(dat[1]);
+		posy = Double.parseDouble(dat[2]);
 		comment = dat[3];
 		poli.clear();
 		bomb.clear();
@@ -23,24 +23,25 @@ public class Dades {
 		for(int i = 4; i < dat.length; i+=3){
 			Staff st = new Staff();
 			st.setId(dat[i]);
-			st.setPosx(Float.parseFloat(dat[i+1]));
-			st.setPosy(Float.parseFloat(dat[i+2]));
+			st.setPosx(Double.parseDouble(dat[i+1]));
+			st.setPosy(Double.parseDouble(dat[i+2]));
 			int n = Integer.parseInt(""+dat[i].charAt(0));
 			if(n == 1)		poli.add(st);
 			else if(n == 2)	bomb.add(st);
 			else if(n == 3)	amb.add(st);
 		}
 	}
+
 	
 	public String getId() {
 		return id;
 	}
 
-	public Float getPosx() {
+	public Double getPosx() {
 		return posx;
 	}
 
-	public Float getPosy() {
+	public Double getPosy() {
 		return posy;
 	}
